pyproject.toml
[project]
name = "sf-full"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp~=3.9.0",
    "annotated-types~=0.7.0",
    "astrapy~=1.5.0",
    "boto3~=1.36.9",
    "cachetools~=5.5.0",
    "cairosvg>=2.8.2",
    "celery>=5.5.3",
    "certifi~=2024.8.30",
    "charset-normalizer~=3.4.0",
    "click~=8.1.7",
    "cloup~=3.0.5",
    "convex>=0.7.0",
    "cython~=3.0.11",
    "decorator~=5.1.1",
    "fastapi[standard]>=0.116.1",
    "ffmpeg-python~=0.2.0",
    "glcontext~=3.0.0",
    "google-ai-generativelanguage~=0.6.10",
    "google-api-core~=2.22.0",
    "google-api-python-client~=2.151.0",
    "google-auth~=2.35.0",
    "google-auth-httplib2~=0.2.0",
    "google-cloud-aiplatform~=1.79.0",
    "google-generativeai~=0.8.3",
    "googleapis-common-protos~=1.65.0",
    "grpcio~=1.67.1",
    "grpcio-status~=1.67.1",
    "gtts~=2.5.3",
    "httplib2~=0.22.0",
    "httpx~=0.27.0",
    "idna~=3.10",
    "imageio-ffmpeg~=0.5.1",
    "inquirer>=3.4.1",
    "isosurfaces~=0.1.2",
    "jinja2~=3.1.4",
    "kokoro-onnx>=0.4.9",
    "krippendorff~=0.8.1",
    "langchain~=0.3.14",
    "langchain-aws>=0.2.13",
    "langchain-community~=0.3.14",
    "langfuse~=2.58.1",
    "langgraph>=0.4.0",
    "langgraph-cli>=0.3.6",
    "litellm~=1.60.5",
    "manim~=0.18.1",
    "manim-chemistry~=0.4.4",
    "manim-circuit~=0.0.3",
    "manim-dsa~=0.2.0",
    "manim-ml~=0.0.24",
    "manim-physics~=0.4.0",
    "manim-voiceover~=0.3.7",
    "manimpango~=0.6.0",
    "mapbox-earcut~=1.0.2",
    "markdown-it-py~=3.0.0",
    "mcp~=1.0.0",
    "mdurl~=0.1.2",
    "moderngl~=5.12.0",
    "moviepy~=2.1.2",
    "multipledispatch~=1.0.0",
    "mutagen~=1.47.0",
    "networkx~=3.4.2",
    "numpy~=2.2.2",
    "openai~=1.61.0",
    "opencv-python~=4.11.0",
    "passlib[bcrypt]~=1.7.4",
    "pillow>=10.4.0",
    "proto-plus~=1.25.0",
    "protobuf~=5.28.3",
    "pyasn1~=0.6.1",
    "pyasn1-modules~=0.4.1",
    "pyaudio~=0.2.14",
    "pycairo~=1.27.0",
    "pydantic~=2.9.2",
    "pydantic-core~=2.23.4",
    "pydantic-settings>=2.4.0",
    "pydub~=0.25.1",
    "pyglet~=2.0.18",
    "pygments~=2.18.0",
    "pylatexenc~=2.10",
    "pyparsing~=3.2.0",
    "pyrr~=0.10.3",
    "pysrt>=1.1.2",
    "python-dotenv~=0.21.1",
    "python-jose[cryptography]~=3.3.0",
    "python-multipart~=0.0.12",
    "python-slugify~=8.0.4",
    "redis>=6.4.0",
    "requests~=2.32.3",
    "rich~=13.9.3",
    "rsa~=4.9",
    "scipy~=1.14.1",
    "screeninfo~=0.8.1",
    "sentencepiece>=0.2.0",
    "skia-pathops~=0.8.0.post2",
    "slowapi~=0.1.9",
    "soundfile~=0.13.1",
    "sox~=1.5.0",
    "srt~=3.5.3",
    "starlette~=0.41.0",
    "statsmodels~=0.14.4",
    "svgelements~=1.9.6",
    "text-unidecode~=1.3",
    "timm>=1.0.19",
    "typing-extensions~=4.12.2",
    "uritemplate~=4.1.1",
    "urllib3~=2.2.3",
    "uvicorn[standard]~=0.32.0",
    "watchdog~=5.0.3",
    "websockets~=12.0",
    "yt-dlp>=2025.7.21",
]
